OM sinking functions improved -- depth-dependent POM sink speed option
To fit the POM data the optimiser settles on low values of wPOM -- far lower than the reference (Dutkiewicz, 2012) value of 10. But a value of 10 makes it difficult to fit the data because POM clears from the surface layers to quickly. There is now an option to model POM sink speed as depth dependent so that wPOM is lower near the surface -- where greater turbulence may reduce sink speeds...